+++ avr_controlRemoto/src/library/UART/UART.hpp
@@ -5,6 +5,7 @@
 #include <util/delay.h>
 #include <stdlib.h>
 #include <stdio.h>
+
 #define USART_BAUDRATE 9600
 #define UBRR_VALUE  (((F_CPU / (USART_BAUDRATE * 16UL))) - 1)
 class UART{
@@ -14,7 +15,7 @@ class UART{
         uint8_t USART0ReceiveByte();
         void UART_write_txt(char *);
         void USART_Uint8_tWrite(uint8_t);
-        void UART_WriteInt(long);
+        void UART_WriteInt(int);
     private:
          
 };
